WIP User Team Controller

+++ backend/src/main/java/com/taptake/backend/controller/UserTeamController.java
@@ -7,6 +7,7 @@ import com.taptake.backend.model.User;
 
 import com.taptake.backend.model.UserTeam;
 import com.taptake.backend.service.GameService;
+import com.taptake.backend.service.PlayerService;
 import com.taptake.backend.service.UserService;
 import com.taptake.backend.service.UserTeamService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,6 +18,7 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.HashSet;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Optional;
@@ -31,6 +33,9 @@ public class UserTeamController {
 
     @Autowired
     private UserService userService;
+
+    @Autowired
+    private PlayerService playerService;
 
     @Autowired
     private GameService gameService;
@@ -55,9 +60,18 @@ public class UserTeamController {
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
         }
 
+        // A user cannot have multiple teams for the same game
+        List<UserTeam> userTeams = userTeamService.findByUser(user);
+        for (UserTeam team : userTeams) {
+            if (team.getGame() == optionalGame.get()) {
+                return ResponseEntity.status(HttpStatus.CONFLICT).build();
+            }
+        }
+
         UserTeam userTeam = new UserTeam();
         userTeam.setUser(user);
         userTeam.setGame(optionalGame.get());
+        userTeam.setPlayers(new HashSet<>());
 
         userTeam = userTeamService.save(userTeam);
 
@@ -79,6 +93,31 @@ public class UserTeamController {
 
     @DeleteMapping
     public ResponseEntity<Object> delete(@RequestParam String id) {
+
+        Optional<UserTeam> optional_UserTeam = userTeamService.findById(UUID.fromString(id));
+
+        if (optional_UserTeam.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
+
+        UserTeam updateUserTeam = optional_UserTeam.get();
+
+        // Get user
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        if ((authentication instanceof AnonymousAuthenticationToken)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
+        String currentUserName = authentication.getName();
+
+        // If user is authenticated, the service must find it
+        User user = userService.findByEmail(currentUserName).get();
+
+        if (updateUserTeam.getUser() != user) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
         userService.deleteOne(UUID.fromString(id));
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
     }
@@ -95,16 +134,44 @@ public class UserTeamController {
 
         UserTeam updateUserTeam = optional_UserTeam.get();
 
-        if (!updateUserTeam.getGame().getIdJogo().toString().equals(userTeamDTO.getIdJogo())) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+        // Get user
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        if ((authentication instanceof AnonymousAuthenticationToken)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
         }
 
-        if (!updateUserTeam.getUser().getIdUsuario().toString().equals(userTeamDTO.getIdJogo())) {
+        String currentUserName = authentication.getName();
+
+        // If user is authenticated, the service must find it
+        User user = userService.findByEmail(currentUserName).get();
+
+        if (updateUserTeam.getUser() != user) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
+        // Game must be the same
+        if (!updateUserTeam.getGame().getIdJogo().toString().equals(userTeamDTO.getIdJogo())) {
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
         }
 
         if (userTeamDTO.getPlayers().size() > updateUserTeam.getGame().getQuantidadeJogadores()) {
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+        }
+
+        for (Player p : updateUserTeam.getActivePlayers()) {
+            if (!userTeamDTO.getPlayers().contains(p.getIdJogador().toString())) {
+                updateUserTeam.removeActivePlayer(p);
+            }
+        }
+
+        for (String playerid : userTeamDTO.getPlayers()) {
+            Optional<Player> optPlayer = playerService.findById(UUID.fromString(playerid));
+
+            if (!optPlayer.isPresent()) {
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+            }
+            updateUserTeam.addActivePlayer(optPlayer.get());
         }
 
         return ResponseEntity.status(HttpStatus.OK).body(userTeamService.update(updateUserTeam));
